requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.2
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
geographiclib==2.0
geopy==2.3.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
Markdown==3.4.3
packaging==23.2
pillow==10.2.0
PyJWT==2.7.0
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.1.0
